Add an "obscurer window" that is OverrideRedirect and on top of everything.
It's meant to guard against compositors giving up compositing and
releasing their window, as it'll then ensure there's still _something_
on top of the desktop.

This shouldn't be able to do any harm - but nevertheless let's evaluate
this properly yet as I have no reproduction case for this problem yet.