auth_pam_x11: Show "Processing..." message only if PAM didn't print a text before.
This serves to handle cases such as PAM asking the user to authenticate
to an external device by asking the user to do something with
PAM_TEXT_INFO then waiting for the user to do so; we then want to keep
PAM's message on the screen until the action happens.

Of course, we still do show the processing message if the last action
was an entry dialog (e.g. password entry).

Also changes on some systems the events on bad password entry:

- Before: Password:, Processing..., Invalid Password, Processing... (sleeping 2 seconds)
- After: Password:, Processing..., Invalid Password (sleeping 2 seconds)

This gives users more time to read the error message.