Restructure how auth and saver windows work:
- Auth window is still always on top.

- Auth window is now a child of the background window, not of the saver
  window.

- Thus, auth window now overlaps the saver window; thus all overlap
  protections for that window have been removed. In worst case something
  may now overlap the saver that we don't want; our protections still
  kick in if it also overlaps the background window (so there's only an
  issue if a saver is misbehaving).

Now auth helpers can create sibling windows; useful for multi-window
auth (be it for showing a windowed prompt on all screens while letting
the saver shine through, or be it for showing a "taskbar" at the bottom
to allow mouse based actions).

This change is compatible to existing auth helpers.