Move blanking from saver_blank to main.c.
Behaves somewhat better if the value is low; still could not get a zero value
to work, as X11 then still wakes up the screen on key-release events.

Note that after this change, SIGUSR1 is only sent if
XSECURELOCK_SAVER_RESET_ON_AUTH_CLOSE is enabled, and savers
should ideally handle it by soft-resetting the saver (implemented in
saver_xscreensaver only for now); not handling it will cause a hard-reset
of the saver by restarting it from scratch.

Fixes #75.